+++ src/reporter/junit.ts
@@ -11,7 +11,13 @@ interface ReportData {
     failures: number;
     errors: number;
     tests: number;
-    rules: { [rule: string]: string[] };
+    time: number;
+    rules: {
+        [rule: string]: {
+            failureList: string[];
+            time: number;
+        };
+    };
 }
 
 // According to https://en.wikipedia.org/wiki/CDATA#Nesting this is how you escape the CDATA end
@@ -20,8 +26,8 @@ const escapeCDATA = (cdata: string): string => {
     return cdata.replace(/\]\]\>/g, ']]]]><![CDATA[>');
 };
 
-const reportTemplate = (suite: string) => `<?xml version="1.0" encoding="utf-8"?>
-    <testsuites package="stricter">
+const reportTemplate = (suite: string, time: number) => `<?xml version="1.0" encoding="utf-8"?>
+    <testsuites package="stricter" time="${time}">
         ${suite}
     </testsuites>`;
 
@@ -37,12 +43,16 @@ const suiteTemplate = (
     ${testcases.join(EOL)}
 </testsuite>`;
 
-const testcaseTemplate = (name: string, failures: string) => `<testcase name="${xmlEscape(name)}">
+const testcaseTemplate = (
+    name: string,
+    failures: string,
+    time: number,
+) => `<testcase name="${xmlEscape(name)}" time="${time}">
     ${failures}
 </testcase>`;
 
-const testcaseFailure = (type: Level, message: string, detail: string = '') =>
-    `<failure type="${type}"${message !== '' ? ` message="${xmlEscape(message)}"` : ''}>${
+const testcaseFailure = (type: Level, detail: string) =>
+    `<failure type="${type}">${
         detail !== ''
             ? `<![CDATA[
 ${escapeCDATA(detail)}]]>`
@@ -54,41 +64,45 @@ export default (report: RuleToRuleApplicationResult): void => {
         (acc, [rule, applicationResult]) => {
             const failureList: string[] = [];
 
-            if (applicationResult.errors) {
+            if (applicationResult.errors.length) {
                 acc.failures += applicationResult.errors.length;
                 acc.errors += applicationResult.errors.length;
-                failureList.push(
-                    testcaseFailure(Level.ERROR, '', applicationResult.errors.join(EOL)),
-                );
+                failureList.push(testcaseFailure(Level.ERROR, applicationResult.errors.join(EOL)));
             }
 
-            if (applicationResult.warnings) {
+            if (applicationResult.warnings.length) {
                 acc.failures += applicationResult.warnings.length;
                 failureList.push(
-                    testcaseFailure(Level.WARNING, '', applicationResult.warnings.join(EOL)),
+                    testcaseFailure(Level.WARNING, applicationResult.warnings.join(EOL)),
                 );
             }
 
-            acc.rules[rule] = failureList;
+            const timeInS = applicationResult.time / 1000;
+            acc.time += timeInS;
+            acc.rules[rule] = {
+                failureList,
+                time: timeInS,
+            };
             acc.tests += 1;
 
             return acc;
         },
-        { failures: 0, errors: 0, tests: 0, rules: {} } as ReportData,
+        { failures: 0, errors: 0, tests: 0, time: 0, rules: {} } as ReportData,
     );
 
     const testcases = Object.keys(data.rules).map(rule =>
-        testcaseTemplate(rule, data.rules[rule].join(EOL)),
+        testcaseTemplate(rule, data.rules[rule].failureList.join(EOL), data.rules[rule].time),
     );
 
     // If we don't have any rules, we still need a dummy testcase so that there's something to parse
     if (testcases.length === 0) {
         data.tests = 1;
-        testcases.push(testcaseTemplate('no errors', ''));
+        testcases.push(testcaseTemplate('no errors', '', 0));
     }
 
     const xml = reportTemplate(
         suiteTemplate('stricter', data.failures, data.errors, data.tests, testcases),
+        data.time,
     );
 
     console.log(xml);